refactor message code and add exampleturns element
