+++ src/components/interfaces/noora-chat/chat/NooraChat.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef, useState } from "react";
 import MessageBox from "./message-box/MessageBox";
 import Header from "./Header";
-import Messages from "./Messages";
+import Messages from "./message-window/Messages";
 import { messageToSpeechParams } from "../../../../scripts/noora-chat/audio_utils";
 import { textToSpeech } from "../../speech/SpeechSynthesizer";
