+++ src/components/interfaces/noora-chat/chat/message-window/ExampleTurns.tsx
@@ -1,0 +1,5 @@
+import React from "react"
+
+export default function ExampleTurns() {
+    return <div></div>
+}
